+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/VideoFile.kt
@@ -1,11 +1,10 @@
 package dev.inmo.tgbotapi.types.files
 
 import dev.inmo.tgbotapi.requests.abstracts.FileId
-import dev.inmo.tgbotapi.types.FileUniqueId
+import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.InputMedia.InputMediaVideo
 import dev.inmo.tgbotapi.types.ParseMode.HTMLParseMode
 import dev.inmo.tgbotapi.types.ParseMode.ParseMode
-import dev.inmo.tgbotapi.types.fileUniqueIdField
 import dev.inmo.tgbotapi.types.files.abstracts.*
 import dev.inmo.tgbotapi.utils.MimeType
 import dev.inmo.tgbotapi.utils.toHtmlCaptions
@@ -18,15 +17,21 @@ data class VideoFile(
     override val fileId: FileId,
     @SerialName(fileUniqueIdField)
     override val fileUniqueId: FileUniqueId,
+    @SerialName(widthField)
     override val width: Int,
+    @SerialName(heightField)
     override val height: Int,
+    @SerialName(durationField)
     override val duration: Long? = null,
+    @SerialName(thumbField)
     override val thumb: PhotoSize? = null,
+    @SerialName(fileNameField)
+    override val fileName: String? = null,
     @SerialName(mimeTypeField)
     override val mimeType: MimeType? = null,
     @SerialName(fileSizeField)
     override val fileSize: Long? = null
-) : TelegramMediaFile, MimedMediaFile, ThumbedMediaFile, PlayableMediaFile, SizedMediaFile
+) : TelegramMediaFile, CustomNamedMediaFile, MimedMediaFile, ThumbedMediaFile, PlayableMediaFile, SizedMediaFile
 
 @Suppress("NOTHING_TO_INLINE")
 inline fun VideoFile.toInputMediaVideo(